Adding cleaning.sql
This file creates the clean versions of the three basic catalogues: 2MASS, VVV and SIRIUS. That is, it merges the duplicated sources. In general, it takes the average position as new position, and the weighted by errors average of the photometry as new photometry (If one of the two sources has no photometry in that band, then only the observed one is taken).
This first commit only adds 2MASS for testing.