Changing the crossmatch to improve 2MASS assignations
Many 2MASS sources are not being matched based on 2MASS wrongly? estimated magnitudes (these do not have errors). This happen to be far from the measured magnitude in the other catalogue. Instead, we make is so that the crossmatch with 2MASS uses only the measured magnitudes. This will affect ~ 100k sources, most of which only have one magnitude observed with 2MASS and are going to be matched based solely on that (on position <1", which is below 2MASS effective angular resolution anyway)