Testing new way to coalesce duplicates
In some case, the duplicated source splits in more than 2 sources, making it hard to merge them. In this attempt, after the initial merge in pairs and then by source_id, the table is xmatch with itself and a unique pair_id is assign to sources that are within xmatch dist. of each other: a list of unique source_ids shared by both entries.