Improved cross-matched with Gaia
This is still not perfect, but at least it takes the best neighbour for all Gaia sources that DO NOT have already a cross-match given by Gaia. Problems: a) the best neighbour is just the one that is closest to the Gaia source. b) 2MASS duplicated sources that have been merged might not have the right sourceId and therefore might not be picked up by the Gaia official crossmatch.